+++ argopy/data_fetchers/erddap_data.py
@@ -23,7 +23,7 @@ from argopy.utilities import list_standard_variables, Chunker, format_oneline
 from argopy.stores import httpstore
 from ..errors import ErddapServerError
 from aiohttp import ClientResponseError
-
+import logging
 
 # Load erddapy according to available version (breaking changes in v0.8.0)
 try:
@@ -37,10 +37,12 @@ except:  # noqa: E722
     # Soon ! https://github.com/ioos/erddapy
 
 
+log = logging.getLogger("argopy.erddap.data")
+
 access_points = ['wmo', 'box']
 exit_formats = ['xarray']
 dataset_ids = ['phy', 'ref', 'bgc']  # First is default
-api_server = 'https://erddap.ifremer.fr/erddap/'  # API root url
+api_server = OPTIONS['erddap']  # API root url
 api_server_check = api_server + '/info/ArgoFloats/index.json'  # URL to check if the API is alive
 
 
@@ -116,10 +118,10 @@ class ErddapArgoDataFetcher(ArgoDataFetcherProto):
             Erddap request time out in seconds. Set to OPTIONS['api_timeout'] by default.
         """
         timeout = OPTIONS["api_timeout"] if api_timeout == 0 else api_timeout
-        self.fs = httpstore(cache=cache, cachedir=cachedir, timeout=timeout, size_policy='head')
+        self.fs = kwargs['fs'] if 'fs' in kwargs else httpstore(cache=cache, cachedir=cachedir, timeout=timeout, size_policy='head')
         self.definition = "Ifremer erddap Argo data fetcher"
         self.dataset_id = OPTIONS["dataset"] if ds == "" else ds
-        self.server = api_server
+        self.server = kwargs['server'] if 'server' in kwargs else OPTIONS['erddap']
 
         if not isinstance(parallel, bool):
             parallel_method = parallel
@@ -141,9 +143,20 @@ class ErddapArgoDataFetcher(ArgoDataFetcherProto):
     def __repr__(self):
         summary = ["<datafetcher.erddap>"]
         summary.append("Name: %s" % self.definition)
-        summary.append("API: %s" % api_server)
+        summary.append("API: %s" % self.server)
         summary.append("Domain: %s" % format_oneline(self.cname()))
         return "\n".join(summary)
+
+    @property
+    def server(self):
+        return self._server
+
+    @server.setter
+    def server(self, value):
+        self._server = value
+        if hasattr(self, 'erddap') and self.erddap.server != value:
+            log.debug("The erddap server has been modified, updating internal data")
+            self._init_erddapy()
 
     def _add_attributes(self, this):  # noqa: C901
         """ Add variables attributes not return by erddap requests (csv)
@@ -247,7 +260,7 @@ class ErddapArgoDataFetcher(ArgoDataFetcherProto):
 
     def _init_erddapy(self):
         # Init erddapy
-        self.erddap = ERDDAP(server=self.server, protocol="tabledap")
+        self.erddap = ERDDAP(server=str(self.server), protocol="tabledap")
         self.erddap.response = (
             "nc"  # This is a major change in v0.4, we used to work with csv files
         )
@@ -337,52 +350,6 @@ class ErddapArgoDataFetcher(ArgoDataFetcherProto):
 
         return vlist
 
-    @property
-    def _dtype(self):
-        """ Return a dictionary of data types for each variable requested to erddap in the minimal vlist """
-        dref = {
-            "data_mode": object,
-            "latitude": np.float64,
-            "longitude": np.float64,
-            "position_qc": np.int64,
-            "time": object,
-            "time_qc": np.int64,
-            "direction": object,
-            "platform_number": np.int64,
-            "config_mission_number": np.int64,
-            "vertical_sampling_scheme": object,
-            "cycle_number": np.int64,
-            "pres": np.float64,
-            "temp": np.float64,
-            "psal": np.float64,
-            "doxy": np.float64,
-            "pres_qc": np.int64,
-            "temp_qc": object,
-            "psal_qc": object,
-            "doxy_qc": object,
-            "pres_adjusted": np.float64,
-            "temp_adjusted": np.float64,
-            "psal_adjusted": np.float64,
-            "doxy_adjusted": np.float64,
-            "pres_adjusted_qc": object,
-            "temp_adjusted_qc": object,
-            "psal_adjusted_qc": object,
-            "doxy_adjusted_qc": object,
-            "pres_adjusted_error": np.float64,
-            "temp_adjusted_error": np.float64,
-            "psal_adjusted_error": np.float64,
-            "doxy_adjusted_error": np.float64,
-            "ptmp": np.float64,
-        }
-        plist = self._minimal_vlist
-        response = {}
-        for p in plist:
-            if p in dref:
-                response[p] = dref[p]
-            else:
-                response[p] = object
-        return response
-
     def cname(self):
         """ Return a unique string defining the constraints """
         return self._cname()
@@ -440,10 +407,9 @@ class ErddapArgoDataFetcher(ArgoDataFetcherProto):
     @property
     def N_POINTS(self) -> int:
         """ Number of measurements expected to be returned by a request """
+        url = self.get_url().replace("." + self.erddap.response, ".ncHeader")
         try:
-            url = self.get_url().replace("." + self.erddap.response, ".ncHeader")
-            with self.fs.open(url) as of:
-                ncHeader = of.read().decode("utf-8")
+            ncHeader = str(self.fs.fs.cat_file(url))
             lines = [line for line in ncHeader.splitlines() if "row = " in line][0]
             return int(lines.split("=")[1].split(";")[0])
         except Exception:
@@ -452,24 +418,28 @@ class ErddapArgoDataFetcher(ArgoDataFetcherProto):
     def to_xarray(self, errors: str = 'ignore'):  # noqa: C901
         """ Load Argo data and return a xarray.DataSet """
 
+        URI = self.uri  # Call it once
+
         # Download data
         if not self.parallel:
-            if len(self.uri) == 1:
+            if len(URI) == 1:
                 try:
-                    ds = self.fs.open_dataset(self.uri[0])
+                    # log.debug("to_xarray: %s" % URI[0])
+                    # log_argopy_callerstack()
+                    ds = self.fs.open_dataset(URI[0])
                 except ClientResponseError as e:
                     raise ErddapServerError(e.message)
             else:
                 try:
                     ds = self.fs.open_mfdataset(
-                        self.uri, method="sequential", progress=self.progress, errors=errors
+                        URI, method="sequential", progress=self.progress, errors=errors
                     )
                 except ClientResponseError as e:
                     raise ErddapServerError(e.message)
         else:
             try:
                 ds = self.fs.open_mfdataset(
-                    self.uri, method=self.parallel_method, progress=self.progress, errors=errors
+                    URI, method=self.parallel_method, progress=self.progress, errors=errors
                 )
             except ClientResponseError as e:
                 raise ErddapServerError(e.message)
@@ -507,7 +477,7 @@ class ErddapArgoDataFetcher(ArgoDataFetcherProto):
         ds.attrs["Fetched_by"] = getpass.getuser()
         ds.attrs["Fetched_date"] = pd.to_datetime("now", utc=True).strftime("%Y/%m/%d")
         ds.attrs["Fetched_constraints"] = self.cname()
-        ds.attrs["Fetched_uri"] = self.uri
+        ds.attrs["Fetched_uri"] = URI
         ds = ds[np.sort(ds.data_vars)]
 
         #
@@ -602,7 +572,7 @@ class Fetch_wmo(ErddapArgoDataFetcher):
         for wmos in wmo_grps:
             urls.append(
                 Fetch_wmo(
-                    WMO=wmos, CYC=self.CYC, ds=self.dataset_id, parallel=False
+                    WMO=wmos, CYC=self.CYC, ds=self.dataset_id, parallel=False, fs=self.fs, server=self.server,
                 ).get_url()
             )
         return urls
@@ -664,5 +634,5 @@ class Fetch_box(ErddapArgoDataFetcher):
             boxes = self.Chunker.fit_transform()
             urls = []
             for box in boxes:
-                urls.append(Fetch_box(box=box, ds=self.dataset_id).get_url())
+                urls.append(Fetch_box(box=box, ds=self.dataset_id, fs=self.fs, server=self.server).get_url())
             return urls