+++ argopy/tests/helpers/utils.py
@@ -79,15 +79,12 @@ has_erddap, requires_erddap = _connectskip(
     "erddap" in AVAILABLE_SOURCES, "erddap data fetcher"
 )
 
-if CONNECTED and has_erddap:
-    log.debug("Check which Erddap dataset are available (eg: core, bgc, ref, index)")
-    try:
-        DSEXISTS = erddap_ds_exists(ds="ArgoFloats")
-        DSEXISTS_bgc = erddap_ds_exists(ds="ArgoFloats-synthetic-BGC")
-        DSEXISTS_ref = erddap_ds_exists(ds="ArgoFloats-ref")
-        DSEXISTS_index = erddap_ds_exists(ds="ArgoFloats-index")
-    except:
-        log.debug("Cannot determine which erddap dataset are available, will skip corresponding tests.")
+if CONNECTED:
+    log.debug("Checking which Erddap dataset are available (eg: core, bgc, ref, index)")
+    DSEXISTS = erddap_ds_exists(ds="ArgoFloats")
+    DSEXISTS_bgc = erddap_ds_exists(ds="ArgoFloats-bio")
+    DSEXISTS_ref = erddap_ds_exists(ds="ArgoFloats-ref")
+    DSEXISTS_index = erddap_ds_exists(ds="ArgoFloats-index")
 else:
     DSEXISTS = False
     DSEXISTS_bgc = False